# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "camp-backend",
    "description": "\n    # AIエージェントアプリケーション API\n\n    LangChain/LangGraphを使用したAIエージェントアプリケーションのバックエンドAPI。\n\n    ## 認証モード: development\n\n    ### 開発モード認証（モック）\n\n    ## 主な機能\n\n    - **AIエージェントとのチャット**: LangGraphによる高度な会話エンジン\n    - **セッション管理**: 会話履歴の保存と取得\n    - **ファイル処理**: ファイルのアップロード・ダウンロード\n    - **認証・認可**: JWT認証 とロールベース制御\n    - **モニタリング**: Prometheusメトリクス収集\n\n    ## 認証\n\n    ほとんどのエンドポイントはゲストアクセス可能ですが、一部の機能では認証が必要です。\n    認証が必要な場合は、`Authorization: Bearer <token>` ヘッダーを含めてください。\n\n    ## レート制限\n\n    APIリクエストは100リクエスト/分に制限されています。\n    ",
    "contact": {
      "name": "開発チーム",
      "email": "nomura.koshiro@gmail.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.2"
  },
  "paths": {
...
    "/api/v1/driver-tree/nodes": {
      "post": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Create Node",
        "description": "ノードを作成します。\n\nArgs:\n    node_data: ノード作成リクエスト\n    current_user: 現在のユーザー\n    driver_tree_service: Driver Treeサービス\n\nReturns:\n    NodeResponse: 作成されたノード",
        "operationId": "create_node_api_v1_driver_tree_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverTreeNodeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverTreeNodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/driver-tree/nodes/{node_id}": {
      "get": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Get Node",
        "description": "ノードを取得します。\n\nArgs:\n    node_id: ノードID\n    current_user: 現在のユーザー\n    driver_tree_service: Driver Treeサービス\n\nReturns:\n    NodeResponse: ノード",
        "operationId": "get_node_api_v1_driver_tree_nodes__node_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverTreeNodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Update Node",
        "description": "ノードを更新します。\n\nArgs:\n    node_id: ノードID\n    node_data: ノード更新リクエスト\n    current_user: 現在のユーザー\n    driver_tree_service: Driver Treeサービス\n\nReturns:\n    NodeResponse: 更新されたノード",
        "operationId": "update_node_api_v1_driver_tree_nodes__node_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverTreeNodeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverTreeNodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/driver-tree/trees": {
      "post": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Create Tree From Formulas",
        "description": "数式からツリーを作成します。\n\nArgs:\n    formula_data: 数式作成リクエスト\n    current_user: 現在のユーザー\n    driver_tree_service: Driver Treeサービス\n\nReturns:\n    TreeResponse: 作成されたツリー",
        "operationId": "create_tree_from_formulas_api_v1_driver_tree_trees_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverTreeFormulaCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverTreeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/driver-tree/trees/{tree_id}": {
      "get": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Get Tree",
        "description": "ツリーを取得します。\n\nArgs:\n    tree_id: ツリーID\n    current_user: 現在のユーザー\n    driver_tree_service: Driver Treeサービス\n\nReturns:\n    TreeResponse: ツリー",
        "operationId": "get_tree_api_v1_driver_tree_trees__tree_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverTreeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/driver-tree/categories": {
      "get": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Get Categories",
        "description": "カテゴリー一覧を取得します。\n\nArgs:\n    current_user: 現在のユーザー\n    driver_tree_service: Driver Treeサービス\n\nReturns:\n    dict: カテゴリーの辞書",
        "operationId": "get_categories_api_v1_driver_tree_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "type": "object"
                  },
                  "type": "object",
                  "title": "Response Get Categories Api V1 Driver Tree Categories Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/driver-tree/kpis": {
      "get": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Get Kpis",
        "description": "KPI一覧を取得します。\n\nArgs:\n    current_user: 現在のユーザー\n\nReturns:\n    KPIListResponse: KPI一覧",
        "operationId": "get_kpis_api_v1_driver_tree_kpis_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverTreeKPIListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/driver-tree/formulas": {
      "get": {
        "tags": [
          "driver-tree"
        ],
        "summary": "Get Formulas",
        "description": "指定されたツリータイプとKPIの数式を取得します。\n\nArgs:\n    current_user: 現在のユーザー\n    tree_type: ツリータイプ\n    kpi: KPI名\n    driver_tree_service: Driver Treeサービス\n\nReturns:\n    FormulaResponse: 数式レスポンス",
        "operationId": "get_formulas_api_v1_driver_tree_formulas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ツリータイプ",
              "title": "Tree Type"
            },
            "description": "ツリータイプ"
          },
          {
            "name": "kpi",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "KPI名",
              "title": "Kpi"
            },
            "description": "KPI名"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverTreeFormulaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriverTreeFormulaCreateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "ツリー名"
          },
          "formulas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Formulas",
            "description": "数式のリスト"
          }
        },
        "type": "object",
        "required": [
          "formulas"
        ],
        "title": "DriverTreeFormulaCreateRequest",
        "description": "数式からツリー作成リクエスト。"
      },
      "DriverTreeFormulaResponse": {
        "properties": {
          "formulas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Formulas",
            "description": "数式のリスト"
          }
        },
        "type": "object",
        "required": [
          "formulas"
        ],
        "title": "DriverTreeFormulaResponse",
        "description": "数式レスポンス。"
      },
      "DriverTreeKPIListResponse": {
        "properties": {
          "kpis": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kpis",
            "description": "KPI名のリスト"
          }
        },
        "type": "object",
        "required": [
          "kpis"
        ],
        "title": "DriverTreeKPIListResponse",
        "description": "KPI一覧レスポンス。"
      },
      "DriverTreeNodeCreate": {
        "properties": {
          "tree_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tree Id",
            "description": "所属するツリーのID"
          },
          "label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Label",
            "description": "ノードのラベル"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "親ノードID（Noneの場合はルートノード）"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator",
            "description": "親ノードとの演算子"
          },
          "x": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "X",
            "description": "X座標"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Y",
            "description": "Y座標"
          }
        },
        "type": "object",
        "required": [
          "tree_id",
          "label"
        ],
        "title": "DriverTreeNodeCreate",
        "description": "ノード作成リクエスト。\n\nAttributes:\n    tree_id: 所属するツリーのID\n    label: ノードのラベル（KPI名や計算要素名）\n    parent_id: 親ノードID（Noneの場合はルートノード）\n    operator: 親ノードとの演算子（+, -, *, /, %など）\n    x: X座標（ツリー表示用）\n    y: Y座標（ツリー表示用）"
      },
      "DriverTreeNodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "ノードID"
          },
          "tree_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tree Id",
            "description": "所属するツリーのID"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "ノードのラベル"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "親ノードID"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator",
            "description": "演算子"
          },
          "x": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X",
            "description": "X座標"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y",
            "description": "Y座標"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/DriverTreeNodeResponse"
            },
            "type": "array",
            "title": "Children",
            "description": "子ノードのリスト"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tree_id",
          "label"
        ],
        "title": "DriverTreeNodeResponse",
        "description": "ノードレスポンス（木構造）。"
      },
      "DriverTreeNodeUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "ノードのラベル"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "親ノードID"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator",
            "description": "演算子"
          },
          "x": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "X",
            "description": "X座標"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Y",
            "description": "Y座標"
          }
        },
        "type": "object",
        "title": "DriverTreeNodeUpdate",
        "description": "ノード更新リクエスト。"
      },
      "DriverTreeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "ツリーID"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "ツリー名"
          },
          "root_node_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Node Id",
            "description": "ルートノードID"
          },
          "root_node": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverTreeNodeResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "ルートノード（木構造全体を含む）"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DriverTreeResponse",
        "description": "ツリーレスポンス。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}